Get docker CgroupsPath from proc when missing
